pyproject.toml
[tool.poetry]
name = "ecgs-service"
version = "0.1.0"
description = "ECGs service"
authors = ["Txema Bermudez <txema.bermudez@gmail.com>"]
readme = "README.md"
packages = [{ include = "ecgs_service" }]

[tool.poetry.dependencies]
python = "3.10.8"
fastapi = "0.85.1"
pytest = "7.1.3"
sqlmodel = "0.0.8"
psycopg2-binary = "2.9.4"
alembic = "1.8.1"
python-dateutil = "2.8.2"
requests = "2.28.1"
uvicorn = "0.19.0"
email-validator = "1.3.0"
python-jose = "3.3.0"
passlib = { extras = ["bcrypt"], version = "1.7.4" }
python-multipart = "0.0.5"


[tool.poetry.group.dev.dependencies]
pytest = "7.1.3"
SQLAlchemy-Utils = "^0.38.3"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
